Apply focused-input preservation to NumberField and add e2e tests
Generalize useTextInputProps to support non-string value types and add a
NumberField wrapper that uses it, so ui.number_field preserves user input
while focused (matching ui.text_field and ui.text_area).

Move the e2e test app and spec from ui_text_field to ui_events and add
coverage for text_area and number_field.
