+++ plugins/ui/src/js/src/widget/WidgetUtils.tsx
@@ -23,7 +23,6 @@ import {
   ListActionGroup,
   ListActionMenu,
   MenuTrigger,
-  NumberField,
   Section,
   Switch,
   TabList,
@@ -87,6 +86,7 @@ import {
   Markdown,
   Menu,
   Meter,
+  NumberField,
   Picker,
   ProgressBar,
   ProgressCircle,
