+++ tests/app.d/tests.app
@@ -20,4 +20,4 @@ file_13=ui_nested_dashboard.py
 file_14=ui_query_params.py
 file_15=ui_home_screen.py
 file_16=ui_routing.py
-file_17=ui_text_field.py
+file_17=ui_events.py
